refactor: move TimeRange to models
