+++ tskv/src/iterator.rs
@@ -19,7 +19,7 @@ use minivec::MiniVec;
 use models::predicate::domain::{ColumnDomains, Domain, PredicateRef, Range, ValueEntry};
 use models::schema::{ColumnType, TableColumn, TskvTableSchema, TIME_FIELD, TIME_FIELD_NAME};
 use models::utils::{min_num, unite_id};
-use models::{FieldId, SeriesId, ValueType};
+use models::{FieldId, SeriesId, TimeRange, ValueType};
 use parking_lot::RwLock;
 use snafu::ResultExt;
 use tokio::time::Instant;
@@ -28,7 +28,7 @@ use trace::{debug, error, info};
 use super::engine::EngineRef;
 use super::error::IndexErrSnafu;
 use super::memcache::DataType;
-use super::tseries_family::{ColumnFile, SuperVersion, TimeRange};
+use super::tseries_family::{ColumnFile, SuperVersion};
 use super::tsm::{BlockMetaIterator, DataBlock, TsmReader};
 use super::{error, ColumnFileId, Error};
 use crate::compute::count::count_column_non_null_values;
