+++ tskv/src/kvcore.rs
@@ -10,7 +10,7 @@ use models::codec::Encoding;
 use models::predicate::domain::ColumnDomains;
 use models::schema::{make_owner, DatabaseSchema, TableColumn};
 use models::utils::unite_id;
-use models::{ColumnId, SeriesId, SeriesKey};
+use models::{ColumnId, SeriesId, SeriesKey, TimeRange};
 use protos::kv_service::{WritePointsRequest, WritePointsResponse};
 use protos::models as fb_models;
 use protos::models_helper::get_db_from_fb_points;
@@ -33,7 +33,7 @@ use crate::index::{ts_index, IndexResult};
 use crate::kv_option::{Options, StorageOptions};
 use crate::schema::error::SchemaError;
 use crate::summary::{Summary, SummaryProcessor, SummaryTask, VersionEdit};
-use crate::tseries_family::{SuperVersion, TimeRange, TseriesFamily};
+use crate::tseries_family::{SuperVersion, TseriesFamily};
 use crate::tsm::codec::get_str_codec;
 use crate::version_set::VersionSet;
 use crate::wal::{WalEntryType, WalManager, WalTask};
